feat: move store out of pages
