+++ src/components/layout/SearchPanel.tsx
@@ -1,6 +1,6 @@
 import { Box, Center, FlexProps, Heading, HStack, Spacer } from "@chakra-ui/react";
 import { observer } from "mobx-react";
-import gameState, { callUpdateGlobalSlot, callUpdateLandfillPile, callUpdateMarketDeck, callUpdatePlayers } from "pages/store";
+import gameState, { callUpdateGlobalSlot, callUpdateLandfillPile, callUpdateMarketDeck, callUpdatePlayers } from "src/store/store";
 import React from "react";
 import PlayingCard from "../PlayingCard";
 import { Search } from "src/entities/carboncityzero/carbonCityZeroPlayer";
